# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gift Multiverse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  /** @nullable */
  coverColor?: string | null;
  publishedAt: Date;
}