# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gift Multiverse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  /** @nullable */
  username?: string | null;
}